feat: ChatState 타입 정의를 state.py로 이동하여 코드 구조 개선
